Fail when requesting multiple_values to yajl backend
The yajl backend has never supported multiple values, and therefore has
never even exposed the option in its basic_parse routines. This commit
adds the option to align the basic_parse signature with the rest of the
backends, but immediately raises an error if multiple_values is
requested. A unit test is included to make sure this behavior works.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>