Add new, enriched entry-level functions
Until now the basic set of functions offered by ijson took always a file
object and returned a generator. If users wanted to use an async file
object they would need to invoke the respective *_async function, which
returns an asynchronous iterable, but otherwise looks identical from an
invocation perspective.  Because the invocation is identical in these
two cases, it makes sense to have a higher-level implementation of these
"entry-level" functions that inspects what the source object is (is it a
file? is it an async file?) and internally invokes and returns the
corresponding lower-level function.

This commit introduces these new enriched entry-level functions. They
are able to idenfity both asynchronous and synchronous file objects,
dispatching the call to the corresponding backend function. Because they
handle file objects they actually replace the old "basic_parse",
"parse", "items" and "kvitems" functions in the backend, providing users
with a simpler calling convention. The old generators have been renamed
to *_gen in case users want to be explicit about calling those.

Note that the yajl2_c backend needed its *_basecoro functions to be put
back as the new generic, enriched entry-level functions rely on the
*_basecoro functions being defined. This was not strictly needed because
common.enrich_backend() will add missing *_basecoro functions, but on
the other hand we would not be getting the performance given by the C
backend's own function. In general it might be good to work in the
future to align the _yajl2_c extension's functions with the rest of the
backends' and expose them directly (instead of having this small salad
layer in the yajl2_c module), we'll have a think about that.

In addition to supporting both flavours of file objects, the new
functions also support iterables, which are assumed to be events
generated from previous steps in the pipeline. In other words, they come
to replace the old ijson.common.{parse,items,kvitems} functions, which
will be marked as deprecated and eventually removed. Note that
basic_parse does not support this extra type of input as its input is
always raw JSON data.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>