Document extended usage of enriched functions
On the one hand we allow async objects now, so this was a simple change
of phrasing. On the other hand though we now have completely new
functionality as we incorporated the ijson.common.* routines
functionalities into the new enriched functions, so this required a
complete new section in the documentation (the ijson.common.* routines
were undocumented in the README previously).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>