Simulate more complex read() function
Adding an extra asyncio.sleep() call and awaiting on it effectively
yields the execution of the coroutine, turning it into a more realistic
test and example. The previous simpler version masked a problem we had
in our C backend, and therefore this change should prevent that
particular problem from creeping in again.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>