Await for utils35._get_read(f) properly
Our C code was not ready to deal with suspended executions of
utils35._get_read, as the logic assumed that once an awaitable was
created to get the read function, it would yield the result (via
StopIteration) after the first next() invocation. This broken logic
meant that when our asynchronous iterable was called again, it would
invoke _get_read() again with no arguments, resulting in an exception
being thrown.

In our unit tests our dummy async read functions don't suspend their
execution, which means that when calling the _get_read(f) function from
C we have always gotten away with this bug.

This patch fixes this problem, correctly iterating as long as necessary
on this first awaitable, rather than only once.

This is the original problem reported in #32.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>