Deal with @types.coroutine awaitables
Awaitable objects that don't have an __await__ method are probably
generators that have been marked with @types.coroutine, and therefore we
should be able to use them directly instead. This check needs only to be
done when invoking the read() function, as we know our _get_read
function is a native coroutine.

What we are *not* yet considering are C-defined python types that have
an tp_as_async.am_await field defined. For the time being we are
assuming those are rare, but it might bite us in the future.

This is the second problem found during the investigation of #32.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>