Generalize async test infrastructure
We are planning to add other reader classes to test the async
infrastructure against (to ensure that our code works correctly against
@type.coroutine functions), and therefore we need to refactor the code
to be reusable in that sense.

Here we separate the code we'll reuse into the _test_async_common.py
module, which _test_async.py uses to define its get_first/get_all
methods based on its own reader.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>