Move yajl patching process to setup.py
When building ijson in yajl embedded mode (IJSON_EMBED_YAJL=1) we
previously had to manually get a copy of yajl, and move files around so
yajl and ijson could be compiled together into a single shared library.
This made it less clear that these two processes had to be performed,
making the setup process a bit more difficult to maintain.

This commit brings the patching process of yajl into setup.py. Instead
of getting a copy of yajl via git clone we use the new git submodule as
the yajl sources, which we first copy into a writable location, and then
modify them as needed. This means that users who want to build an ijson
package in yajl embedded mode only need to ensure that the submodule is
present, and that the environment variable is set.

This commit also updates the AppVeyor pipeline to use this new
procedure.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>