Fix 32-bit behavior on cffi backends
The cffi backend, when running in python 2.7 and being use_float=True,
always yielded long() objects, even if the values were small,
breaking the current set of expectations of our tests. With this change
the values returned by this backend will be int() if they are small
enough, of long() if they go beyond the underlying native int C type,
when running on python 2.7.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>