Fix yajl2 backend integer callback declaration
The two ctypes-based yajl backends (for versions 1 and 2, respectivelly)
used the same definition for their integer callback, while in reality
this is different between the two different major versions of yajl:
version 1 of yajl gives values to the callback using a "long" integer
type, while in yajl 2 this value is passed as a "long long" instead. Our
callback erroneously declared "long" as the value type for both major
versions.

This change fixes the declaration of the integer callback, making it
depend on the major version of yajl being wrapped by ctypes. When yajl
version 2 is wrapped, the callback now declares its values to be passed
with a "long long" type as they should. At the same time, an extra
int(v) call is made on the value to force an "int" type when possible in
python 2.7, rather than yielding "long" python objects otherwise.

Together with this change, I'm also simplifying a bit the setup
infrastructure for these callbacks to avoid some code duplication.

Depending on the platform and architecture, this change is significant,
because "long" is "at least" 32 bits in size (but it's 64 bits in x64_64
Linux/MacOS), while "long long" is defined to be "at least" 64 bits in
size. This affects the range of values that can be represented by yajl
itself, and therefore passed to our library.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>