Clarify parse_value's state_stack's state
The previous version of this code had 3 lines missing from the coverage.
These lines were missing because they were after an "if state_stack"
condition, which I initially thought could be true or false depending on
the program execution. Extensive testing has showed us that these lines
are never touched, and closer code inspection also reveals that the
state_stack list is never empty before entering into the main "if/elif"
control block, as this is ensured in the previous try/catch block.

This commit therefore removes the conditions resulting in missing lines,
and instead adds an assert expression after the try/catch block to show
more explicitly the state of the execution.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>