Create builder_t objects as part of their parents
We previously created the builder objects with a separate memory
allocation call. This process has a chance to fail, although in
our case it is very minimal. This small possibility however accounted
for some code that was missing coverage (because the error never
occurred). Incidentally, we also used raw memory allocation calls
instead of the python-specific ones, which has the potential of bringing
trouble.

This commit changes how builder_t objects are managed, allocating its
memory together with the object that contains it (ItemsBasecoro and
KVItemsBasecoro). This removes an extra memory allocation from our code,
removing the need to check the result of that extra memory allocation,
and therefore allowing us to remove this otherwise uncovered code.

This change should also have some further performance improvements, as
an extra memory allocation has been removed, and given that the
builder's memory now resides next to the rest of its container object.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>