Remove memory leaks in C backend
The C backend had a couple of memory leaks in the initialization
routines of some of the generators it implements. Because these leaks
occurred when generators were initialized, they did not show up easily,
unless many such objects were created. This is unlike memory leaks in
the parsing process, which are apparent with a full pass over a single
parsing operation.

Two of these leaks were solved by adding the corresponding decrease in
the reference count of the objects leaked. The rest were solved by
improving how we built tuples out of sequence slices, changing our
previous manual, memory-leaking approach to use a simpler GetSlice
operation, which is easier to ready anyway.

This commit addresses #34.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>